[Google Kickstart] 2021 Round B
Solved all of Problem A and half of problem C. Was exteremly close on
both of the other two. TLE on the second (although I know I was close
to a better one-pass solution) and an MLE on problem D. Need to figure
that out since I was super close to solving that.